Remove torch audio part, add example and test
